Extract QueryDatabasePathFactory.
