+++ firebase-web/src/test/java/io/spine/web/firebase/query/FirebaseQueryBridgeTest.java
@@ -99,7 +99,7 @@ class FirebaseQueryBridgeTest {
         Query query = queryFactory.all(Empty.class);
         QueryProcessingResult result = bridge.send(nonTransactionalQuery(query));
 
-        assertThat(result, instanceOf(FirebaseQueryProcessingResult.class));
+        assertThat(result, instanceOf(FirebaseQueryResult.class));
     }
 
     @Test
